Deletion race condition handled

+++ pkg/entities/sjf/service/delete.go
@@ -2,7 +2,9 @@ package service
 
 import (
 	"context"
+	"fmt"
 
+	"github.com/DaZZler12/sjf-be/pkg/entities/sjf/constants"
 	"go.mongodb.org/mongo-driver/bson"
 )
 
@@ -10,9 +12,12 @@ func (sjfService *SJFService) Delete(ctx context.Context, filters *bson.M) error
 	if filters == nil {
 		filters = &bson.M{}
 	}
-	_, err := sjfService.Get(ctx, filters)
+	obtainedSJFJob, err := sjfService.Get(ctx, filters)
 	if err != nil {
 		return err
+	}
+	if obtainedSJFJob.Status == constants.Running || obtainedSJFJob.Status == constants.Pending {
+		return fmt.Errorf("cannot delete a job that is in %s state", obtainedSJFJob.Status)
 	}
 	return sjfService.sjfStore.Delete(ctx, filters)
 }
